feat: refactor Kafka and OPA container imports, enhance logging, and improve container management
