+++ tests/fixtures/policy_stores.py
@@ -3,7 +3,7 @@ from images import cedar_image, opa_image
 from testcontainers.core.network import Network
 
 from tests.containers.cedar_container import CedarContainer
-from tests.containers.opa_container import OpaContainer, OpaSettings
+from tests.containers.OPA.opa_container import OpaContainer, OpaSettings
 from tests.containers.settings.cedar_settings import CedarSettings
